BJX2 core, console print, earlier sanity checks.

+++ jx2vm/jx2i_opc_arith.c
@@ -1148,7 +1148,7 @@ void BJX2_Op_SHAD_RegReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-			ctx->regs[op->rn]<<(shl&31);
+			((s32)ctx->regs[op->rn])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
@@ -1164,7 +1164,7 @@ void BJX2_Op_SHLD_RegReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-			ctx->regs[op->rn]<<(shl&31);
+			((u32)ctx->regs[op->rn])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
@@ -1333,8 +1333,8 @@ void BJX2_Op_SHAD_RegImmReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-//			((s32)ctx->regs[op->rm])<<(shl&31);
-			(ctx->regs[op->rm])<<(shl&31);
+			((s32)ctx->regs[op->rm])<<(shl&31);
+//			(ctx->regs[op->rm])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
@@ -1350,8 +1350,8 @@ void BJX2_Op_SHLD_RegImmReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-//			((u32)ctx->regs[op->rm])<<(shl&31);
-			(ctx->regs[op->rm])<<(shl&31);
+			((u32)ctx->regs[op->rm])<<(shl&31);
+//			(ctx->regs[op->rm])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
@@ -1399,8 +1399,8 @@ void BJX2_Op_SHAD_RegRegReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-//			((s32)ctx->regs[op->rm])<<(shl&31);
-			(ctx->regs[op->rm])<<(shl&31);
+			((s32)ctx->regs[op->rm])<<(shl&31);
+//			(ctx->regs[op->rm])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
@@ -1416,8 +1416,8 @@ void BJX2_Op_SHLD_RegRegReg(BJX2_Context *ctx, BJX2_Opcode *op)
 	if(shl>=0)
 	{
 		ctx->regs[op->rn]=
-//			((u32)ctx->regs[op->rm])<<(shl&31);
-			(ctx->regs[op->rm])<<(shl&31);
+			((u32)ctx->regs[op->rm])<<(shl&31);
+//			(ctx->regs[op->rm])<<(shl&31);
 	}else
 	{
 		ctx->regs[op->rn]=
